+++ src/ElectronNET.Host/api/webContents.js
@@ -111,6 +111,13 @@ module.exports = (socket) => {
             electronSocket.emit("webContents-domReady" + id);
         });
     });
+    socket.on("register-webContents-zoomChanged", (id) => {
+        const browserWindow = getWindowById(id);
+        browserWindow.webContents.removeAllListeners("zoom-changed");
+        browserWindow.webContents.on("zoom-changed", (_, zoomDirection) => {
+            electronSocket.emit("webContents-zoomChanged" + id, zoomDirection);
+        });
+    });
     socket.on("webContents-openDevTools", (id, options) => {
         if (options) {
             getWindowById(id).webContents.openDevTools(options);