+++ Changelog.md
@@ -3,6 +3,7 @@
 ## ElectronNET.Core
 
 - Updated dependencies
+- Updated `WebContents` zoom level APIs to use `double` instead of `int` (#956)
 - Improved migration checks to honor custom output paths and to detect `ProjectGuid` in publish profiles (#946)
 - Fixed single instance handling on macOS (#1040)
 - Fixed slow socket bridge startup by binding to an explicit loopback address (#1103)
@@ -11,6 +12,7 @@
 - Fixed cross-compilation behavior on same platform (#1098) @epsnm
 - Fixed resolution of unrelated target (#1099) @epsnm
 - Fixed false alarm for `ELECTRON001` on a root `package-lock.json` (#946)
+- Added `WebContents.OnZoomChanged` event (#956)
 - Added target framework customization (#1095) @epsnm
 - Added configurable Electron root directory for custom packaging layouts (#1106) @DYH1319
 - Added ability for `custom_main.js` to modify command line switches (#1029) @AeonSake